refactor: extract TimeGrid from DesktopLayout
Extract the main calendar time grid (hour lines, GTD frame zones, task
blocks with conflict layout, drag/resize handles, inline editing,
routine pills) into a TimeGrid component. Shared between desktop and
tablet via context breakpoint values.

DesktopLayout drops from ~2370 to ~806 lines. Massive import cleanup
removes ~40 unused imports that moved into sub-components.
